Rename CSET._workflow_utils to CSET.workflow
It is about to contain the whole workflow, rather than just some
utilities.
