+++ tests/workflow/test_finish_website.py
@@ -17,7 +17,7 @@
 import json
 import logging
 
-from CSET._workflow_utils import finish_website
+from CSET.workflow import finish_website
 
 
 def test_copy_rose_config(monkeypatch, tmp_path):
